refactor: Disunite repositories from cubit directory
Disunite repositories from cubit directory
