+++ courageous_people/lib/log_in/cubit/log_in_cubit.dart
@@ -1,5 +1,5 @@
 import 'package:bloc/bloc.dart';
-import 'package:courageous_people/log_in/cubit/log_in_repository.dart';
+import 'package:courageous_people/log_in/repository/log_in_repository.dart';
 import 'package:courageous_people/log_in/cubit/log_in_state.dart';
 import 'package:flutter/cupertino.dart';
