+++ courageous_people/lib/review/cubit/review_cubit.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:bloc/bloc.dart';
-import 'package:courageous_people/log_in/cubit/log_in_repository.dart';
+import 'package:courageous_people/log_in/repository/log_in_repository.dart';
 import 'package:courageous_people/log_in/cubit/log_in_state.dart';
 import 'package:image_picker/image_picker.dart';
